refactor: move code to features folders
