+++ packages/api-reference/src/features/MultipleDocuments/DocumentSelector.vue
