+++ packages/api-reference/src/features/MultipleDocuments/index.ts
@@ -1,1 +1,2 @@
 export { default as DocumentSelector } from './DocumentSelector.vue'
+export { useMultipleDocuments } from './useMultipleDocuments'
